test improvements and better metrics
